# This is a combination of 3 commits. # The first commit's message is: # This is a combination of 2 commits. # The first commit's message is:
inityan commin for mrp

# The 2nd commit message will be skipped:

#	commin temp work to test in fedrogesa

# The 2nd commit message will be skipped:

#	mrp

# The 3rd commit message will be skipped:

#	moduling pv

+++ point_of_sale/point_of_sale.py
@@ -484,28 +484,9 @@ class pos_order(osv.osv):
         order_ids = []
         for tmp_order in orders:
             order = tmp_order['data']
-            partner_id = order['partner_id'] if 'partner_id' in order else False
-            if 'order_id' in order:          
-                order_id = order['order_id']
-                if order_id != []:
-                    self.write(cr, uid, order_id, {
-                        'user_id': order['user_id'] or False,
-                        'session_id': order['pos_session_id'],                    
-                        'pos_reference':order['name'],
-                        'partner_id':partner_id,                    
-                        }, context)              
-                    self.update_lines(cr, uid, order_id, order['lines'], context)
-                else: print "!!!!Error Perdida de id de la orden En actulizacion de productos MRP¡¡¡¡¡¡¡¡¡¡"
-            else:
-                order_id = self.create(cr, uid, {
-                    'name': order['name'],
-                    'user_id': order['user_id'] or False,
-                    'session_id': order['pos_session_id'],
-                    'lines': order['lines'],
-                    'pos_reference':order['name'],
-                    'partner_id':partner_id,
-                }, context)          
-
+            if 'order_id' in order: order_id=self.update_sales(cr, uid, order, context)
+            else: order_id=self.create_sales(cr, uid, order, context)
+                
             for payments in order['statement_ids']:
                 payment = payments[2]
                 self.add_payment(cr, uid, order_id, {
@@ -515,7 +496,7 @@ class pos_order(osv.osv):
                     'payment_name': payment.get('note', False),
                     'journal': payment['journal_id']
                 }, context=context)
-
+    
             if order['amount_return'] > 0:
                 session = self.pool.get('pos.session').browse(cr, uid, order['pos_session_id'], context=context)
                 cash_journal = session.cash_journal_id
@@ -532,17 +513,47 @@ class pos_order(osv.osv):
                     'payment_name': _('return'),
                     'journal': cash_journal.id,
                 }, context=context)
-            order_ids.append(order_id)
-            self.signal_paid(cr, uid, order_ids)
+            order_ids.append(order_id)            
+            self.signal_paid(cr, uid, order_id)
         return order_ids
-
+    
+    def update_sales(self, cr, uid, order, context=None):
+        partner_id = order['partner_id'] if 'partner_id' in order else False
+        order_id = order['order_id']
+        if order_id != []:
+                   self.write(cr, uid, order_id, {
+                       'user_id': order['user_id'] or False,
+                       'session_id': order['pos_session_id'],                    
+                       'pos_reference':order['name'],
+                       'partner_id':partner_id,                    
+                       }, context)              
+                   self.update_lines(cr, uid, order_id, order['lines'], context)
+        else:
+            print "!!!!Error Perdida de id de la orden En actualizacion de productos¡¡¡¡¡¡¡¡¡¡"
+            print "!!ORDEN ACTUAL : ", tmp_order,"¡¡¡"
+            return False
+        return order_id
+    
+    def create_sales(self, cr, uid, order, context=None):
+        order_id = self.create(cr, uid, {
+                    'name': order['name'],
+                    'user_id': order['user_id'] or False,
+                    'session_id': order['pos_session_id'],
+                    'lines': order['lines'],
+                    'pos_reference':order['name'],                    
+                }, context)          
+        temp= self.pool.get('pos.order').browse(cr,uid,order_id)        
+        self.update_lines(cr, uid, temp.id, context)
+        return order_id
+    
     def get_dic(self,seq, key):
         return dict((d[key], dict(d, index=index)) for (index, d) in enumerate(seq))
     
     def get_current_lines(self,order_lines):
-        #calcula ordenes actuales        
+        #calcula ordenes actuales              
         current_lines=[]
         for line1 in order_lines:
+            print line1
             d= {                            
                         'discount':line1.discount,
                         'price_unit':line1.price_unit,
@@ -553,50 +564,49 @@ class pos_order(osv.osv):
             current_lines.append(d)
         return current_lines
         
-    def update_lines(self, cr, uid, order_id, lines, context = None):
+    def update_lines(self, cr, uid, order_id, lines = None, context = None):
         
         #TODO optimizar actualizar,añador y remover lineas de productos
         #ordenes nuevas
-        new_lines = []
-        for line in lines:
-            new_lines.append(line[2])        
-        dic_new_lines = self.get_dic(new_lines,'product_id')
-        
-        order_lines = self.pool.get('pos.order').browse(cr, uid, order_id, context).lines
-        current_lines=self.get_current_lines(order_lines)
-        dic_current = self.get_dic(current_lines,'product_id')
-        
-               
-        print "Inicio: "        
-        
-        print "Dicionario de diccionarios: ",dic_current
-        
-        #conjuntos eliminacion de ordenes que ya no estan en las nuevas ordes        
-        x = set(dic_new_lines)       
-        y = set(dic_current)
-        z = y-x
-        for element in z:            
-            a = dic_current[element]
-            print "Elemento a eliminar: ",a
-            self.pool.get('pos.order.line').unlink(cr, uid,a['id'],context)
-      
-        for lin in new_lines:
-           print "Orden actual : ",current_lines
-           print "Intentando grabar producto : ",lin               
-           if lin['product_id']  not in  dic_current:
-              lin['order_id'] = order_id              
-              print "Agregado linea de producto: ",lin
-              self.pool.get('pos.order.line').create(cr, uid, lin, context)
-           else:
-               print "Registro actulizado en: linea de producto: ",lin
-               self.pool.get('pos.order.line').write(cr, uid,dic_current[lin['product_id']]['id'], {                               
-                          'discount':lin['discount'],
-                          'price_unit':lin['price_unit'],
-                          'qty':lin['qty'],
-                      }, context)       
-        print "Fin" 
-        return "True"   
+        if lines !=None:
+            new_lines = []
+            for line in lines:
+                new_lines.append(line[2])        
+            dic_new_lines = self.get_dic(new_lines,'product_id')
             
+            #if order_id[0]!= [] : order_id = order_id[0]
+            order_lines = self.pool.get('pos.order.line').browse(cr, uid, order_id, context)
+            current_lines=self.get_current_lines(order_lines)            
+            dic_current = self.get_dic(current_lines,'product_id')
+            
+            print "Dicionario de diccionarios: ",dic_current
+            
+            #conjuntos eliminacion de ordenes que ya no estan en las nuevas ordes        
+            x = set(dic_new_lines)       
+            y = set(dic_current)
+            z = y-x
+            for element in z:            
+                a = dic_current[element]
+                print "Elemento a eliminar: ",a
+                self.pool.get('pos.order.line').unlink(cr, uid,a['id'],context)
+          
+            for lin in new_lines:
+               print "Orden actual : ",current_lines
+               print "Intentando grabar producto : ",lin               
+               if lin['product_id']  not in  dic_current:
+                  lin['order_id'] = order_id              
+                  print "Agregado linea de producto: ",lin
+                  self.pool.get('pos.order.line').create(cr, uid, lin, context)
+               else:
+                   print "Registro actulizado en: linea de producto: ",lin
+                   self.pool.get('pos.order.line').write(cr, uid,dic_current[lin['product_id']]['id'], {                               
+                              'discount':lin['discount'],
+                              'price_unit':lin['price_unit'],
+                              'qty':lin['qty'],
+                          }, context)
+        else: return False      
+        return True   
+
     def unlink(self, cr, uid, ids, context=None):
         for rec in self.browse(cr, uid, ids, context=context):
             if rec.state not in ('draft','cancel'):
